Fix endoint types for consistency with the Logz API (#56)
Previously, the endpoint type values were not matching what was returned
from the Logz.io API. This made the consumers of the package to manage
their own constants when using the endpoint types. This commit fixes
that by using the same values returned from the Logz API as the constant
for endpoint type. It also adds a helper function to get the URL for
each endoint type.