Fix header manipulation to Custom Endpoints
Neither all the Custom endpoints have header definition which means that
the provider expected use the values in the header to set headerMap
causing TERRAFORM CRASH when terraform tried to check the current state
of endpoints already existents.

To work around it in the previous version we were forced to set faked
headers to avoid "Out of index error" and this handling purpose to avoid
this inconvenience.